Multivariable Series passes all tests
