+++ symengine/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SRC
     polynomial.cpp
     series.cpp
     series_generic.cpp
+    series_generic_multivariate.cpp
     rings.cpp
     ntheory.cpp
     dense_matrix.cpp
@@ -74,7 +75,7 @@ set(HEADERS
     visitor.h    eval_double.h    diophantine.h cwrapper.h  printer.h  real_double.h
     eval_mpfr.h  eval_arb.h       eval_mpc.h     complex_double.h         series_visitor.h
     real_mpfr.h  complex_mpc.h    type_codes.inc lambda_double.h series.h series_piranha.h
-    basic-methods.inc   series_flint.h  series_generic.h sets.h  derivative.h
+    basic-methods.inc   series_flint.h  series_generic.h series_generic_multivariate.h sets.h  derivative.h
 )
 
 # Configure SymEngine using our CMake options:
