+++ symengine/visitor.h
@@ -22,6 +22,7 @@
 #include <symengine/real_mpfr.h>
 #include <symengine/complex_mpc.h>
 #include <symengine/series_generic.h>
+#include <symengine/series_generic_multivariate.h>
 #include <symengine/series.h>
 #include <symengine/series_piranha.h>
 #include <symengine/series_flint.h>
